naive: set canvas size 700x700. Fix .content positioning.
Canvas was previously 400px wide x 400px high. This sets it to 700px wide by 700px high. Further, we do this in the <canvas> element itself, rather than in CSS.

We also fix a bug in the positioning of the div containing the canvas, .content. Because all the elements in the .heading div were floated, the .content was being positioned without consideration for these .heading elements. We add "clear: both" to .content to fix its position.

(Oddly, the <canvas> was being positioned correctly regardless, but if we added a background color to .sim-container for debugging purposes, we saw it was not positioned at the same location as the <canvas>.)