Naive: fix bug where matrix data wasn't displayed, or loading correctly
The display bug was simply a "display: none" I added during debugging to hide the matrix list. I removed this.

The loading bug was due to the fact that `graph`, which was used as an argument to loadMatrices(), was only set when the promise from setup() was fulfilled. This caused that argument to be invalid. I changed it so that we only call loadMatrices() after the promise from setup() is fulfilled.