Naive: if two points overlap, compute a slightly better force direction
If two points have the exact same coordinates, the direction of the force is (0,0). No matter how much force is applied, the point will not be moved (0 * x = 0).

We previously calculated a janky sort-of random direction for the force in this case. This commit changes the way we do that to be slightly more random.