Naive: change how we calculate alpha to be closer to d3
In d3, alpha (the dampening factor for forces, which increases over time) is calculated as starting at 0.1, and then each iteration sets `alpha *= 0.99`. Our code now does the same thing.

We do make one notable change: we divide stepNumber by TILES_PER_ITERATION and floor that. In this way, stepNumber only seems to increase once we've processed all the tiles in the graph.