`run-site`: add '--dev' mode switch
In when run-site is passed the --dev argument, as in `run-site --dev`, then it will enable "development mode."

In normal (non-dev) mode, the Web Inspector will not be open when launching WebKit.app and opening our page. The script will also not continually relaunch WebKit.app when it closes.

When in dev mode, run-site will open the Web Inspector and continually relaunch WebKit.app, which is how run-site previously worked all the time. Thus, dev mode is equivalent to the current behavior, and regular mode disables some of the current behavior.

The normal mode is useful when giving demos, etc. Any time we're not debugging code, basically.