`run-site`: run server on loopback, not 0.0.0.0
Previously, the run-site script would choose between the `serve` system command (if it was available) or the SimpleHTTPServer Python module. Both of these run the web server with the listening address set to 0.0.0.0:8888, i.e. they listen on all addresses. This meant your local web server was open to the world.

We now include our own Python script embedded in run-site. This script is adapted from my own `serve` command and uses Python to launch a simple HTTP server. The main difference is that we can now choose the listen address. By default in the run-site script, this is 127.0.0.1.

This makes the run-site script more secure and flexible going forward. Yay.