run-site: prompt the user for the path of WebKit
If the `run-site` script can't find the path to WebKit.app from environmental variables or the config file ~/.webkitwebcl, it will use AppleScript to display a prompt to the user to find WebKit.app. Should that succeed, then the path is saved into the config file for future use. If it fails, then the run-site script exits.