Dislike midnodes don't repulse (experimental)
Previously, midnodes of different orders (for example, the first midnode of one edge, and the second midnode of another edge) would repulse each other. However, this isn't in keeping with the original paper's algorithm.

This change causes midnodes of different orders to not interact at all, as the paper lays out. We do this in a very inefficient way, but it suffices for now.

This change arguable does not improve how the simulation looks, so we may want to reverse this commit later...