Use new renderer directly; build config as extern
Now, instead of shimming new renderer calls through the old renderer, main.js uses the new renderer directly.

Because the new renderer is driven by a config file (we already wrote one for SC, named renderer.sc.js), we have Browserify generate this as its own module aliased as "render-config". We also list that module as an external module in the main bundle.

The idea is that we can write a number of rendering config files for different rendering setups, and each is compiled to its own module aliased as render-config. Then, in the HTML page, we include one of those script files, which the main bundle will import and use.