Refactor animation loop, VBO fetching into two fns
Previously, in node-driver.js, create() would create the graph, start the animation loop, and return a signal that produced the VBO data from the graph every n=30ms. server.js then subscribed to this signal to get the VBOs it should emit to the client.

This refactor renames create() to createAnimation(), and it only creates the graph and starts the animation loop. It returns a signal that fires once each animation loop.

fetchData() is a new exported function that takes a graph object, and returns an observable sequence containing one element, the current VBO data (and the number of elements that represents.)

In server.js, we sample the animation signal every time the client sends back a handshake, then we fetch the VBO data and emit that to the client. Even though we're fetching on-demand (vs. proactively,) thus adding the time to fetch the VBOs to the overall latency of a client-server loop, this seems to run faster than the old way. Go figure.