Who has two thumbs and kernels compiling on Linux?
This guy right here.

There turned out to be some incredibly dumb, not-to-spec thing where nVidia's OpenCL compiler in linux only allows geometric-style references to the elements of a vector, not array-style. So, for `foo = (float4)(1,2,3,4)`, `foo.y` is accepted, but `foo[1]` is not. This goes completely against the spec and is stupid and dumb and I wish a pox on the engineers at nVidia right now...

But the graph streamer demo seems to work on AWS!