User can now specify address & port of server.js
This commit modifies server.js to take an optional command-line argument of IP:Port to listen on. Previously, sever.js took an optional argument of the port to listen on, but was hard-coded to bind to `localhost` as its IP.

The default value if no command-line arguments are given to server.js is to listen on localhost:10000. server.js will print it's listening address + port to stdout on startup.

The IP:Port argument is the first (and only) argument server.js takes. The argument may leave out either the IP or Port and server.js will use the default value of the missing part.

Examples: `server.js 192.168.0.10:9000`, `server.js :9000`, `server.js 9000`, `server.js 192.168.0.10`, `server.js localhost:9000`