server: serve compressed data using express server
Instead of creating another server on another port (using the `connect` module instead of the existing `express` module to boot) to listen for XHR requests and respond with compressed data, we now re-use the existing express server to listen for and serve compressed VBOs.

In addition to allowing us to drop the `connect` dependency, the main effect of this is to serve all data from the same port. The request to the server, other than dropping the explicit IP:Port part, should look and act exactly the same as before.