Load StreamGL via npm, rather than absolute path
Hooray, no more needing to set the path to the `nodecl` directory in `config.js` anymore! Now we load StreamGL via npm, and so the path to it should be automatically discovered. Be sure you `npm install` StreamGL, do a `npm link` in the StreamGL directory, and do a `npm link StreamGL` in the `naive` directory (though these should automatically happen when you run `npm install` in StreamGL and then naive.)