Remove Bunyan logging of metrics (fixes #1)
Previously, after sending performance metrics to Boundary (if running
in production mode,) we would also log the performance metrics to
stdout with Bunyan. This created a lot of noise in the terminal and
logs.

This commit removes Bunyan logging of metrics. We can (and likely
should) re-add Bunyan as part of our logging system in the future.