Remove node-webgl shaders; move client WebGL shaders to ./shaders dir
We previously kept around the shaders to the desktop node-webgl version
of the graph viz demo. These were outdated and needing update. We move
the client-side shaders (sent to the client by the graph viz server as
part of the render-config) to the old location of the node-webgl
shaders, './shaders'. Paths have been updates as necessary.