Disable fact gathering by default
This allows us to include machines in the inventory that we can't (yet)
connect to. This is most applicable to changing the sshd port from 22
to 61630. In that case, the inventory lists the (for example)
example.com:61630 as the machine, but our first task is to use the
local command `nc` to test if example.com:61630 is open. If it's not,
we change the Ansible SSH port to example.com:21, connect, change the
sshd config to use port 61630, then change the Ansible SSH port back to
example.com:61630 and continue. With fact gathering on, this wouldn't
work because before we could test for example.com:61630 connectivity
locally, Ansible would try to gather facts about example.com by
connecting to example.com:61630, fail, and the playbook would fail
before we could change Ansible's SSH port.

We can (and should) still gather facts about remote machines at a point
where we know it is safe to connect, by running a task with the `setup`
module.