Automatically change sshd port on machines using port 22
Previously, we'd have to manually edit the Ansible SSH port to port 22,
run the playbook until it changed the sshd port to our custom port, let
it fail when it could no longer connect to port 22, then manually
re-edit the Ansible SSH port to use our custom port, then re-run the
playbook.

We now run a local task to check if we can connect to the custom SSH
port on the remote machine, and if that fails, we automatically set the
Ansible SSH port to 22, change the sshd port on the server, then
automatically reset the Ansible SSH port to our custom one, and
continue the playbook. Sweet.