Preflight: don't set private key permissions; doesn't work
Ansible will throw the error about key permissions before we get a
chance to change them in the preflight role (even if the tasks are all
local, and even trying to do all sorts of work-arounds where we change
the private key file variable, etc.) We really do need to set this
before Ansible runs. So, remove the task from preflight.