Change ansible_managed template var to not include user ID of uploader
While it's nice for files managed by Ansible's templates to include the
user ID of the person who was running Ansible when the file was
uploaded. However, this caused unnecessary re-uploads and reports of
changes by Ansible. The problem is that if two users run Ansible, the
first user's Ansible run will upload the resolved templated file with
their name in it. When the second user runs Ansible, the template will
now contain their name, and Ansible will recognize that the resolved
file is different than the one on disk and re-upload it, even if the
actual template didn't change at all.

Ansible default value for ansible_managed included the date and time of
the upload in the template, which (while useful) is even worse because
the template will *always* be different on *every run*.