Move Restart Supervisor handler to app_server to fix potential bug
When we moved installation of Supervisor from the common role to the
app_server role, we forgot to move the Restart Supervisor handler with
it. This meant that, if we were ever installing/configuring Supervisor
fresh, it may trigger the notification and could have caused an error
because of the missing handler.