Splunkfwd: fix bug which could cause install of Boundary Meter to fail
The Ansible `uri` module has a bug when saving the response, where it
will un-escape escaped quotes (meaning `"\""` turns into `"""`). This
was causing the Boundary Meter install to fail, since the response from
the Boundary API call is a Bash script we need to then run, which has
escaped quotes.

The solution here is to use `curl` (via the Ansible `command` module)
to download the script (which it does without mangling the escapes),
then to run it as normal.