Enable SSH control socket & compression
SSH control socket should be enabled by default in Ansible. However,
because we were overriding `ssh_args` in `ansible.cfg` to enable SSH
Agent forwarding, it was accidentally disabled. Ansible is **much**
faster when it's enabled, so we correct this error.

We also enable compression on the SSH connection, to help with file
transfers.

(The `PermitLocal=no` directive is just there because my local
`~/.ssh/config` has a local command which seems to mess up Ansible when
run. This stops it from being run.)