CUDA: upload Apt preference to prioritize NVIDIA packages over Ubuntu
This fixes a bug where CUDA 7.0 (and specifically the `nvidia-346`
driver) gave an Apt error during install, preventing them from
successfully installing. The cause was that both the NVIDIA and the
Ubuntu apt repos provide the `nvidia-settings` package, and the Ubuntu
version was overriding the NVIDIA version. However the Ubuntu version
is too old to satisfy the requirements of the (NVIDIA repo sourced)
nvidia-346 drivers, and thus the drivers weren't installing because the
dependencies could not be satisfied.

The solution is to upload an Apt preferences file which prioritizes
NVIDIA's apt repo over all others.