app_server: tell logrotate to chown all logfiles to www:graphistry-dev
When logrotate rotated log files, the rotated files would retain their original ownership (www:graphistry-dev); and logrotate would create new, empty logfiles  in place of the rotated files with that same ownership. However, for some reason, when logrotate would compress logfiles (which it does on the second rotation of a logfile and older,) it would have ownership of www:www.

This commit adds a shell command to the logrotate config to `chown` all the files in the `/var/log/graphistry-json/` directory. This command is run after all the files in the directory have been rotated.