workers: have reaper.py ping Boundary using hostname from by Ansible
We previously had reaper.py find it's hostname and FQDN using Python's `socket` module, which would get it from the OS. However, due to the mysteries of Linux networking, this would always come up as `localhost`. Fun times.

This commit updates reaper.py so that its hostname & FQDN are filled in (via template variables) by Ansible, using the hostname found used in its inventory file, when we do an Ansible deploy. This is the same way we set the system hostname, so this should be good to go. Now, why Python wasn't picking up on the system hostname we set is a mystery for the ages.