Remove unused 'packages' directory
This was added four months ago in commit 523bece with message "calc npm deps", then never touched again. So far as I can tell, these files are not used or referenced anywhere at all.

Since the data is redundant with the package.json in the actual package repos, I'm going to delete it for now. If we need it later, we can always reverse this commit :)