Enforce Ansible version 1.9, and document how to fix in `local-dev.md`
Only Ansible 1.9.x is compatible with our playbooks. Older version (1.8.x) and the newest version (2.0.x) are not compatible, and will not be able to run our deploys. Unfortunately, your OS may automatically install or upgrade to Ansible 2.0.x.

The deploy script now checks the version of Ansible it's being run under, and will fail early with an informative message if the `<major>.<minor>` version number isn't 1.9.

This commit also adds documentation to the `documentation/local-dev.md` document warning users about this problem, and explains how to install the proper version of Ansible.