Remove deploy.graphistry.com from inventory
deploy.graphistry.com is our CI server. It was added to our playbook inventory (for both staging and production) because we were (months ago) working to automate setup/deployment of that server. However, work never progressed very far on that task, and the config of the server is currently in flux. Since trying to run Ansible on it was, at best, not benefiting us, this commit removes it from the inventories to save time during deployments (and headaches).