proxy: remove unnecessary 'Restart nginx' notification
We had this notification set to fire after removing one of the components of the old version of nginx. But it doesn't really make sense to restart something we just uninstalled :P This was just a bad copy+paste job.

We still restart nginx after doing other relevant changes to it, however.