Disable retry files in Ansible config (annoying + useless feature)
The "retry files" feature of Ansible makes it so that Ansible saves a list of the hosts that fialed when a playbook run fails. The idea is that you can re-run your playbook, but limit the run to just these failed hosts, instead of redoing all hosts.

This never works for us, because: (1) we only really run a playbook against one host; (2) we need to run some commands on localhost to do a deploy, and limiting the playbook to just the failed host will prevent this; (3) if a playbook fails, it's likely an error that applies to all hosts, not some transient problem on just one host; and (4) it's not even reliable, since the playbook stops at the end of a role that caused the error, meaning the 'working' hosts we might exclude could be stuck in a half-deployed state.

Not only is the future dumb, but it drops the file with the list of failed hosts right in `~/`. So now you have useless files littering your home directory.

This commit disables the feature entirely as a consequence of the above.