cuda: make shell login script that sets CUDA env vars work in sh + BASH
The CUDA role copies a file to `/etc/profile.d` (which is run whenever any user logs in) which sets some environmental variables to allow CUDA/OpenCL to compile/run (e.g., `$LD_LIBRARY_PATH`, etc.).

Previously, this script was only compatible with BASH, not old-school POSIX `sh`. But occasionally, the user's shell could be set to `sh` instead of BASH, which would cause this script to fail, and the login shell to get all weird.

This commit updates the script to work in BASH and `sh`, without changing functionality.