setup.js: link devDependencies in addition to normal dependencies
Previously, when `setup.js --link` was run, it would only examine the `dependencies` field of a package when figuring out what packages it can `npm link <dependency>`. It would ignore the `devDependencies` of the package.

This is a problem, since `npm` *will* install devDependencies of a module when running `npm install` locally. This meant that there were some Graphistry dependencies installed per-package, in `./node_modules`, rather than linked to your git clone of the package, which is not the expected behavior.

This commit changes `setup.js` to examine both `dependencies` and `devDependencies` when figuring out what packages can be `npm link <package>`.