Don't use ASCII color in `npm test` output on Linux
ASCII color codes only really when viewed inside a terminal. If you redirect output of `npm test` to a logfile or other non-terminal destination, the ASCII color codes just add garbage noise to the output. This is particularly a problem for Jenkins log files. Though Jenkins does not directly parse/use the STDOUT of `npm test` for anything, developers might want to look at the log of it to see why a build failed.

We still would like to see ASCII colors when a developer does run `npm test` in their terminal, so we need a compromise.

Since all of our developers use Mac OS X on their machines, and all of our servers use Linux, this commit changes it so that if the test runner (karma) is running on Linux, colors are disabled; otherwise, colors are enabled. This should give is clean STDOUT logs on Jenkins, and nice, colored output for developers locally.