Remove now-unused `proxy` group vars file
"Proxy" is now a role we apply to the `central` server group, rather than being a distinct server group in and of itself. Because of this, setting variables for the `proxy` server group has no effect. Therefore, this commit removes the file that set those variables (none were needed, since the `central` groups variables now cover all the variables we want to set.)