site.yml: move applying `mongo` from own play to existing play
We already have a play that applies to `central` servers, so no reason why we shouldn't apply the `mongo` role to `central` there, rather than in it's own stand-alone play. This also speeds up Ansible a small bit.

There should be no change in behavior from this; we still install MongoDB on `central`, just as before.