site.yml: move `workers` play below `central`; remove `deploy` play
We now run the `central` play, deploying all of our "central"-related stuff, before we run the `workers` play. The reason for this is that `central` now installs MongoDB, so running it first means that Mongo will be alive when the workers start up and begin to ping.

We also remove the `deploy` play, since that isn't being used at the moment (no servers were included as part of the `deploy` server set.)