reaper.py: use CLI arg, rather than ec2 API, for 'hostname' option
Previously, we used the magic ec2 IP `http://169.254.169.254` to query the ec2 metadata API for the instance's public IP, then used that when pinging the pinger database, or querying the DB to find workers running on the instance. We did this because (due to NAT) the public IP of an instance wasn't directly known by the OS, and we used the IP in the database to construct URLs that both our services (central, viz-server) and end users (via StreamGL) made use of.

However, our infrastructure has evolved so that the above method no longer made sense, because:

1. We can no longer assume our apps are running on AWS. That magic ec2 IP doesn't exist on a customer's network, and so we can't use it to find our IP.
2. Even on ec2, we would prefer our apps connect to each other using their private IP. Amazon bills traffic using private IPs as LAN traffic (which costs $0), but bills traffic using public IP addresses as internet traffic (which costs >$0). This is true even if the actual machines communicating are *exactly the same* in both cases, are in the same region, same VPC, etc. If it uses a public IP, AWS' routers always send the traffic out to the internet, then back into AWS, and bill as such.
3. End users no longer connect directly to our web apps. Instead, all user traffic is reverse proxied through nginx to our apps. This means our web apps don't need to know their public IP, or even have one, since it is never given out to the user, and no direct connections are made from the internet to our apps. It is thus sufficient to advertise a private IP/DNS name in the pinger DB.

Given those reasons, we remove the code in reaper.py that queries the ec2 metadata API to determine the server's IP/address to use in pinger DB updates/queries. Instead, the user can manually set the address used for this using the `--address <server address to use>` command line argument to reaper.py. The `viz-server` processes and reaper.py running on the same server should use the same exact address, and this address should be resolvable from `central` to the server.

Since we now only run central/viz-server/reaper.py on a single box, all of our apps are accessible to all of our other apps using the address `localhost` :P As a convenience, we default to using `localhost` as the server address if no `--address` argument is specified to reaper.py on the command line.