reaper.py: remove `hostname` template var; no more templating left
We previously were computing the value of `hostname` via a template variable, filled-in by Ansible during deploy. This variable was only ever used as part of reaper.py's log messages.

This commit removes that template variables, and replaces it with a call to Python's `socket.gethostname()` function to find the hostname at runtime. This function isn't always 100% accurate, but since it's only used in log messages, it's not really a big problem if it's not accurate.

This change removes the last remaining JINJA2 template value in reaper.py (hooray!). reaper.py is now 100% pure python and can be deployed by simply copying to the server, rather than being forced to deploy it via Ansible to convert the template to pure Python.

(The file is still found at `roles/worker/templates/reaper.py.j2`, even though it isn't a template. This is to make the git history easier to read when I make a PR for this branch. I'll move it to the `files` directory, and remove the `.j2` extension after we've merged the PR. It shouldn't cause any problems for now; if Ansible runs it through JINAJ2 during deploy, the same exact file as input will be generated, which is what we want.)

This commit also removes the `reaper_pid` global variable from reaper.py. We only used this variable in one place, in our log output string. Since it's trivial to compute, we just calculate the value in-place in the single place it's needed, and remove the global variable.