worker: in Supervisord config for reaper.py, specify DB url using arg
reaper.py has sane defaults for its program options, and so it can generally be run without having to pass any command line arguments, and everything will work.

The one exception is the DB url, at least when running on our own (Graphistry) infrastructure. You see, the MongoDB URL contains the database server username and password as part of it. This means that, if we wanted our Graphistry MongoDB server used as the default value, we would have to embed our server login credentials into reaper.py. Any customer who had our apps installed could look into reaper.py's source and get our super-secret credentials.

I couldn't bring myself to allow this, and so I made the default DB URL use a made-up username+password. As a result, if we don't specify the correct URL as an argument to reaper.py, it will default to using the made-up username+password, and won't be able to login to our DB server.

This commit changes the Supervisord config used to run reaper.py, so that our real DB URL is passed as an argument to reaper.py, when Supervisor runs it. This should allow reaper.py to run on our infrastructure without trouble.