Always mount the 'data cache' outside the container.
Allow a volume name to be used for the data cache, or use a directory on the host system.
Our deploys each use a different directory, and this will, by default, mount the data cache in a deploy-local directory.
Add notes to the airgapped documentation to inform sysadmins of the utility of this option.