nginx: strip request body from proxied requests to central
Central's 3 jobs, in order of frequency, are: redirect requests intended for workers to those workers; serve static content for the landing page; and log client errors at `/error`. All but that last one don't use the request body, so to save overhead, we strip it from the request before proxying it to central. This saves bandwidth, etc. For `/error`, we define its own `location` block that does not strip the request body.