Nginx: turn off proxy buffering everywhere
We previously only turned it off for `/worker` locations. However, our new "named" worker locations (`@worker10001` and the like) aren't allowed to be nested, per nginx config file syntax. Rather than adding a directive to each of the 32 locations to set proxy buffering to off, we set that option globally.

The effect of this change is that proxy buffering is now off for central, whereas before it was on. However, that doesn't seem like it would be a problem, so let's go for it :)