nginx: fix for config syntax error about `map` not being allowed 'here'
nginx requires that `map` directives be defined in the `http` context. Because of our include file locations/content, we were actually defining it in the `server` context.

This commit puts the `map` directive in the `upstreams-workers.conf` file, which we include into `server.conf` at the `http` context (`routes-workers.conf`, where it was previously defined, is included in the `server` context, by comparison).

This commit also modifies the `generate-routes-workers.sh` script to generate both `upstreams-workers.conf` and `routes-workers.conf`. Previously we only generated the latter, and made the user manually configure the former. But if both files need to change with the number of workers, we might as well generate both. The script also writes those files' content directly to the files, instead of printing to STDOUT and making the user pipe that into the file(s).