nginx: block Slack's link preview bot from accessing our server
Whenever somebody sends a Slack message with a URL in it, Slack’s bot tries to load the page to generate a preview. In our new viz-app architecture, this will cause the bot to requisition a new GPU worker if somebody pastes a viz URL.

This commit blocks user agents with “Slack” in their name, via the nginx config. Not elegant, but Slack has no direct way to turn off their bot (it doesn’t respect `robots.txt` for example). We should figure out a better, more generalized solution in the future.