Remove password protection from viz-app
This affects staging (and halts the rollout to production).

We are currently running bcrypt on each HTTP request for each vbo. As currently implemented, we have a work factor of 12, so we are doing 4096 iterative SHAs to check the password.

There are many possible fixes for this, and once those fixes are in viz-app, we should revert this commit.