Ensure viz-app box override config file is alway created (DEV-1255)
Viz-app’s node-convict config is read from multiple files (where are merged by node-convict): a ‘box override’ config (created when the instance is first configured to run our apps), and a ‘deploy override’ config (created on each Jenkins deploy of our apps).

Previously, we didn’t check to make sure the ’box override’ config existed before we mounted it into the Docker container. If this file doesn’t when `launch.sh` starts the viz-app Docker container, Docker will try to mount the non-existent file (per our Docker command-line options), see that it doesn’t exist, assume that we meant to mount a directory (not a file), and create a new empty directory with the path of the config file (`/var/graphistry/viz-app-config.json`).

viz-app’s config loader code has been updated to check that its config files are files (rather than directories) before trying to load them, which fixes the crash caused by this bug. However, the fact that our config file is a directory (rather than a file) can cause future deployments to fail, if it tries to write a file to that same path.

This commit changes `launch.sh` to ensure that the box override config file is created as an empty JSON file, if it doesn’t exist at launch time. It also checks to see if there’s a directory at that path, and deletes it if so.